faster grouper   - sort on each term of the clause   - binary search all levels deep to quickly find what to group together   - binary search to find the first and last items of the sortedset   - regression test-suite pases   - TODO: cross-check memory leaks

+++ engine/src/errorhandlers.c
@@ -88,7 +88,7 @@ print_trace (void)
 {
 #define BUFSIZE 10
   void *array[BUFSIZE];
-  size_t size = backtrace (array, BUFSIZE);
+  int size = backtrace (array, BUFSIZE);
   char ** strings = backtrace_symbols (array, size);
 
   puts("\nStack Trace: \n");
